Implement live variable analysis
Implement live variable analysis for function parameters, variables and constants.
Define the range from start to end offset where a variable is used and allocate space for used variables. Also define which variables need to be cleared before use.
Move constant values to a lower offset on the stack if possibles.

Signed-off-by: Adam Laszlo Kulcsar <kuladam@inf.u-szeged.hu>